+++ app/src/main/java/com/sopt/now/compose/Data/Service/FriendAuthService.kt
@@ -1,5 +1,6 @@
-package com.sopt.now.compose.Data
+package com.sopt.now.compose.Data.Service
 
+import com.sopt.now.compose.Data.Response.FriendResponseDto
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
